requirements.txt
tablib
Django
django-import-export